backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "broadmail"
version = "0.1.0"
description = "Production-ready mass email platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "arq>=0.25.0",
    "resend>=2.0.0",
    "aiosmtplib>=3.0.0",
    "jinja2>=3.1.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "bleach>=6.1.0",
    "svix>=1.20.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "factory-boy>=3.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
